Refactor link and image handlers
